Completanto el carrito

+++ Tienda/models.py
@@ -18,7 +18,7 @@ class T_Producto(models.Model):
     p_categorias = models.ForeignKey(User, on_delete = models.CASCADE, null = True)
     pro_categorias =  models.ManyToManyField(prueba_Categoria, verbose_name = "Categorias")
     cantidad = models.IntegerField(null = True) 
-    precio = models.DecimalField(max_digits = 8, decimal_places = 3, null = True)
+    precio = models.IntegerField( null = True)
     created = models.DateTimeField(auto_now_add = True)
     updated = models.DateTimeField(auto_now_add = True)
